update WordPress Plugin: Really Simple CAPTCHA

+++ blog/wordpress/wp-content/plugins/really-simple-captcha/really-simple-captcha.php
@@ -6,10 +6,10 @@ Description: Really Simple CAPTCHA is a CAPTCHA module intended to be called fro
 Author: Takayuki Miyoshi
 Author URI: https://ideasilo.wordpress.com/
 Text Domain: really-simple-captcha
-Version: 2.0
+Version: 2.0.1
 */
 
-define( 'REALLYSIMPLECAPTCHA_VERSION', '2.0' );
+define( 'REALLYSIMPLECAPTCHA_VERSION', '2.0.1' );
 
 class ReallySimpleCaptcha {
 
@@ -254,7 +254,10 @@ class ReallySimpleCaptcha {
 				}
 
 				$file = $this->normalize_path( $dir . $filename );
-				$stat = stat( $file );
+
+				if ( ! file_exists( $file ) || ! ( $stat = stat( $file ) ) ) {
+					continue;
+				}
 
 				if ( ( $stat['mtime'] + $minutes * 60 ) < time() ) {
 					if ( ! unlink( $file ) ) {
